address-space: split out layout into dedicated source file
Split out layout into dedicated source file.
